+++ tests/test_elements.py
@@ -449,3 +449,15 @@ def test_element_dirty_name_warning(sanitize_name):
         with warnings.catch_warnings():
             warnings.simplefilter("error", DirtyNameWarning)
             _ = cheetah.Marker(name=dirty_name, sanitize_name=sanitize_name)
+
+
+@pytest.mark.for_every_element("element")
+def test_element_no_internal_dirty_name_warning(element):
+    """Test that no internal `DirtyNameWarning` are produced by methods of `Element`."""
+    element.name = "dirty:element"
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("error", category=DirtyNameWarning)
+
+        _ = element.clone()
+        _ = element.split(torch.tensor(1.0))